datacollection/requirements.txt
requests>=2.32.3
python-dotenv>=1.0.0
azure-storage-blob>=12.17.0
